Tested Functionality for 2D Lighting + Glow Effects
Created a limited test environment in Oliver_Test_Scene to test 2D lighting and Glow effects that could be used in final project. Added in an additional package of test assets to use as environment within test scene. Added an emission map for one of the test assets (LightBandit Sprite Atlas). Added a barebones controller script to test Emissions map with various animations and character movement.
*Note script and sprites are for testing purposes only and are not intended to be used in the further main builds