Removed and Renamed Scenes to reduce clutter. Added in parallax effect to environment and added collision volumes
Removed a majority of unused scenes and renamed most other so that it was clear what they were for. Added in parallax to LevelDemoScene through the creation of the ParallaxController script. Also added collision volumes to the scene so that the player couldn't fall off the "front" or "back" of the level.