Added in Final Level 1 Scene
Added in the final level 1 scene from Louise and began work on porting over all the systems from the previous version of level 1. Finished creating the Tank Animations and Animation Controller. Cleaned up all the enemy Animation Controllers. Changed out the Tank Prefab Sprite for the new Sprite. Changed the Glow Color of the Summoner to be more readable against the Tank