Collider Fixes and Continuing Frontend/UI work
Fixed bounds collider issues within the level and added in colliders to all objects that needed them. Added in the gate mover scripts and slightly changed the position of the level transition bounds. Continued working on the in level UI by adding in new sprites. Added in cutscene images to project in preparation for adding them to the opening and closing screens