Continued Scene Polishing
Added all SFX to the project, also added in the loading screen assets. made adjustments to the scene lighting due to differences between build and editor scenes. Updated the Input systems to allow for controllers to be used in the menu scenes. Added in functionality to trigger the pause menu with a controller. Added in code to prevent enemies from triggering the camera switcher component. Changed navmesh agent size for baking, adjusted the navmesh bake. Added in cutscene images to the menu scenes.