Continued Scene Polishing
Added in functionality to detect if the player is using a controller and change UI text that is displayed. Added in SFX to Level1 scene and all objects within it that needed SFX. Added in Player lives and all associated functionality, including loosing lives, gaining lives via score and enabling the Restart UI menu when the player has no lives left. Added in the main menu background. Added in tutorial popup at the beginning of the level. Added in explosive barrel to tutorial area.